+++ source/main/zettlr-ipc.js
@@ -302,10 +302,10 @@ class ZettlrIPC {
         this._app.openQL(cnt)
         return true
 
-      // DEBUG TODO
       case 'switch-theme-berlin':
       case 'switch-theme-bielefeld':
       case 'switch-theme-frankfurt':
+      case 'switch-theme-karl-marx-stadt':
         // Set the theme accordingly
         global.config.set('display.theme', cmd.substr(13))
         break